WIP - extract URLSession call into HTTP layer
